+++ Tests/Functional/Configuration/RegistrationFormConfigurationCheckTest.php
@@ -6,24 +6,23 @@ namespace OliverKlee\Seminars\Tests\Functional\Configuration;
 
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
 use OliverKlee\Oelib\Configuration\DummyConfiguration;
-use OliverKlee\Seminars\Configuration\SharedConfigurationCheck;
+use OliverKlee\Seminars\Configuration\RegistrationFormConfigurationCheck;
 
 /**
- * @covers \OliverKlee\Seminars\Configuration\SharedConfigurationCheck
+ * @covers \OliverKlee\Seminars\Configuration\RegistrationFormConfigurationCheck
  */
-final class SharedConfigurationCheckTest extends FunctionalTestCase
+final class RegistrationFormConfigurationCheckTest extends FunctionalTestCase
 {
     /**
      * @var array<int, string>
      */
     protected $testExtensionsToLoad = [
-        'typo3conf/ext/static_info_tables',
         'typo3conf/ext/oelib',
         'typo3conf/ext/seminars',
     ];
 
     /**
-     * @var SharedConfigurationCheck
+     * @var RegistrationFormConfigurationCheck
      */
     private $subject;
 
@@ -38,7 +37,7 @@ final class SharedConfigurationCheckTest extends FunctionalTestCase
 
         $this->configuration = new DummyConfiguration();
 
-        $this->subject = new SharedConfigurationCheck($this->configuration, 'plugin.tx_seminars');
+        $this->subject = new RegistrationFormConfigurationCheck($this->configuration, 'plugin.tx_seminars_pi1');
     }
 
     /**
@@ -60,8 +59,8 @@ final class SharedConfigurationCheckTest extends FunctionalTestCase
         $this->subject->check();
 
         $result = $this->subject->getWarningsAsHtml();
-        self::assertArrayHasKey(0, $result);
-        self::assertStringContainsString('plugin.tx_seminars', $result[0]);
+        self::assertArrayHasKey(1, $result);
+        self::assertStringContainsString('plugin.tx_seminars_pi1', $result[1]);
     }
 
     /**
@@ -74,7 +73,21 @@ final class SharedConfigurationCheckTest extends FunctionalTestCase
         $this->subject->check();
 
         $result = $this->subject->getWarningsAsHtml();
-        self::assertArrayHasKey(0, $result);
-        self::assertStringContainsString('plugin.tx_seminars', $result[0]);
+        self::assertArrayHasKey(1, $result);
+        self::assertStringContainsString('plugin.tx_seminars_pi1', $result[1]);
+    }
+
+    /**
+     * @test
+     */
+    public function checkWithAdditionalAttendeesEnabledCreatesErrors(): void
+    {
+        $this->configuration->setAsBoolean('createAdditionalAttendeesAsFrontEndUsers', true);
+
+        $this->subject->check();
+
+        $result = $this->subject->getWarningsAsHtml();
+        self::assertArrayHasKey(1, $result);
+        self::assertStringContainsString('plugin.tx_seminars_pi1', $result[1]);
     }
 }